refactor: re-organize TaskCompletionSource
